0.1.5, finish 2nd phase(lists, tags, objectives, annotations, import igdb, ) and init 3rd and final

+++ app/src/main/java/com/example/taurusgamevault/Model/room/entities/TagGame.kt
@@ -7,6 +7,6 @@ import androidx.room.Entity
     primaryKeys = ["game_id", "tag_id"]
 )
 data class TagGame(
-    val game_id: Int,
-    val tag_id: Int
+    val game_id: Long,
+    val tag_id: Long
 )
